refactor: extract scanner Base struct, unify tool registration, split console.go
- Type Deps.Logger as telemetry.Logger instead of any for compile-time safety
- Extract toolargs.Base with shared Logger/Proxy/WorkDir fields and methods,
  embed in all 7 scanner tool Commands to eliminate repeated boilerplate
- Create per-tool register.go for neutron/spray/zombie/gogo, simplify the
  central register_command.go to only handle the composite scan command
- Split console.go (1331→848 lines) into keybindings.go and banner.go
