+++ pkg/tui/console.go
@@ -8,7 +8,6 @@ import (
 	"io"
 	"os"
 	"path/filepath"
-	"sort"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -21,10 +20,8 @@ import (
 	"github.com/chainreactors/aiscan/pkg/agent"
 	ioaclient "github.com/chainreactors/ioa/client"
 	"github.com/reeflective/console"
-	"github.com/reeflective/readline/inputrc"
 	rlterm "github.com/reeflective/readline/terminal"
 	"github.com/spf13/cobra"
-	"golang.org/x/term"
 )
 
 const agentPromptCommandName = "__prompt"
@@ -132,225 +129,6 @@ func NewAgentConsoleWithTerminal(ctx context.Context, option *cfg.Option, appInf
 	menu.Command = repl.rootCommand()
 	c.SwitchMenu("agent")
 	return repl
-}
-
-func configureAgentReadline(c *console.Console) {
-	if c == nil {
-		return
-	}
-	shell := c.Shell()
-	cfg := shell.Config
-	_ = cfg.Set("autocomplete", true)
-	_ = cfg.Set("usage-hint-always", false)
-	_ = cfg.Set("history-autosuggest", true)
-	_ = cfg.Set("show-all-if-ambiguous", true)
-	_ = cfg.Set("show-all-if-unmodified", true)
-	_ = cfg.Set("menu-complete-display-prefix", true)
-	_ = cfg.Set("page-completions", false)
-	_ = cfg.Set("completion-query-items", 1000)
-	_ = cfg.Set("bell-style", "none")
-	_ = cfg.Set("enable-bracketed-paste", false)
-	// Bind Tab to menu-complete so arrow keys navigate the dropdown.
-	for _, keymap := range []string{"emacs", "emacs-standard", "vi-insert"} {
-		_ = cfg.Bind(keymap, `\t`, "menu-complete", false)
-		_ = cfg.Bind(keymap, inputrc.Unescape(`\e[Z`), "menu-complete-backward", false)
-	}
-}
-
-func (r *AgentConsole) configureInterruptKey() {
-	if r == nil || r.console == nil || r.console.Shell() == nil {
-		return
-	}
-	shell := r.console.Shell()
-	shell.Keymap.Register(map[string]func(){
-		agentConsoleInterruptCommandName: func() {
-			r.handleEscapeInterruptKey()
-		},
-	})
-	escape := inputrc.Unescape(`\e`)
-	for _, keymap := range []string{"emacs", "emacs-standard"} {
-		_ = shell.Config.Bind(keymap, escape, agentConsoleInterruptCommandName, false)
-	}
-}
-
-func (r *AgentConsole) configureCtrlCKey() {
-	if r == nil || r.console == nil || r.console.Shell() == nil {
-		return
-	}
-	shell := r.console.Shell()
-	shell.Keymap.Register(map[string]func(){
-		agentConsoleCtrlCCommandName: func() {
-			r.handleCtrlC()
-		},
-	})
-	ctrlC := inputrc.Unescape(`\C-c`)
-	for _, keymap := range []string{"emacs", "emacs-standard"} {
-		_ = shell.Config.Bind(keymap, ctrlC, agentConsoleCtrlCCommandName, false)
-	}
-}
-
-func (r *AgentConsole) handleCtrlC() {
-	if r.InterruptCurrentRun() {
-		return
-	}
-	if r.pendingExit.Load() {
-		os.Exit(0)
-	}
-	r.pendingExit.Store(true)
-	fmt.Fprintf(r.stderr, " Press Ctrl+C again to exit\n")
-	go func() {
-		time.Sleep(3 * time.Second)
-		r.pendingExit.Store(false)
-	}()
-	shell := r.console.Shell()
-	shell.Display.AcceptLine()
-	shell.History.Accept(false, false, errors.New(os.Interrupt.String()))
-}
-
-func (r *AgentConsole) configureVerbosityToggleKey() {
-	if r == nil || r.console == nil || r.console.Shell() == nil {
-		return
-	}
-	shell := r.console.Shell()
-	shell.Keymap.Register(map[string]func(){
-		agentConsoleToggleVerbosityCommandName: func() {
-			r.handleToggleVerbosity()
-		},
-	})
-	ctrlO := inputrc.Unescape(`\C-o`)
-	for _, keymap := range []string{"emacs", "emacs-standard"} {
-		_ = shell.Config.Bind(keymap, ctrlO, agentConsoleToggleVerbosityCommandName, false)
-	}
-}
-
-func (r *AgentConsole) handleToggleVerbosity() {
-	out := r.ensureOutput()
-	if out == nil {
-		return
-	}
-	current := out.VerbosityLevel()
-	next := (current + 1) % 3
-	out.SetVerbosity(next)
-	label := out.VerbosityLabel()
-	if out.color.Enabled {
-		fmt.Fprintf(r.stderr, "\n%s %s\n",
-			out.dim("verbosity:"),
-			out.color.Wrap(label, outputpkg.ANSICyan))
-	} else {
-		fmt.Fprintf(r.stderr, "\nverbosity: %s\n", label)
-	}
-}
-
-func (r *AgentConsole) handleEscapeInterruptKey() {
-	if r == nil || r.console == nil || r.console.Shell() == nil {
-		return
-	}
-	shell := r.console.Shell()
-	pending := string(shell.Keys.Read())
-	if pending == "" {
-		pending = readPendingTerminalBytes(agentConsoleEscapeSequenceWait)
-	}
-	keymap := string(shell.Keymap.Main())
-	if feed, ok := agentConsoleEscapeSequenceFeed(shell.Config.Binds[keymap], pending); ok {
-		shell.Keys.Feed(true, []rune(feed)...)
-		return
-	}
-	if pending != "" {
-		shell.Keys.Feed(true, []rune(pending)...)
-		return
-	}
-	r.InterruptCurrentRun()
-}
-
-func agentConsoleEscapeSequenceFeed(binds map[string]inputrc.Bind, pending string) (string, bool) {
-	if len(binds) == 0 || pending == "" {
-		return "", false
-	}
-	sequence := inputrc.Unescape(`\e`) + pending
-	matches := make([]string, 0, 4)
-	for seq := range binds {
-		readlineSeq := agentConsoleReadlineSequence(seq)
-		if len(readlineSeq) <= 1 || !strings.HasPrefix(readlineSeq, inputrc.Unescape(`\e`)) {
-			continue
-		}
-		if strings.HasPrefix(sequence, readlineSeq) {
-			matches = append(matches, seq)
-		}
-	}
-	sort.Slice(matches, func(i, j int) bool {
-		left := agentConsoleReadlineSequence(matches[i])
-		right := agentConsoleReadlineSequence(matches[j])
-		if len(left) == len(right) {
-			return left < right
-		}
-		return len(left) > len(right)
-	})
-	for _, seq := range matches {
-		bind := binds[seq]
-		replacement, ok := agentConsoleEquivalentNonEscapeBind(binds, bind)
-		if !ok {
-			continue
-		}
-		return replacement + sequence[len(agentConsoleReadlineSequence(seq)):], true
-	}
-	return "", false
-}
-
-func agentConsoleEquivalentNonEscapeBind(binds map[string]inputrc.Bind, target inputrc.Bind) (string, bool) {
-	if target.Action == "" {
-		return "", false
-	}
-	candidates := make([]string, 0, 4)
-	for seq, bind := range binds {
-		if bind.Action != target.Action || bind.Macro != target.Macro || strings.HasPrefix(agentConsoleReadlineSequence(seq), inputrc.Unescape(`\e`)) {
-			continue
-		}
-		candidates = append(candidates, seq)
-	}
-	sort.Slice(candidates, func(i, j int) bool {
-		if len(candidates[i]) == len(candidates[j]) {
-			return candidates[i] < candidates[j]
-		}
-		return len(candidates[i]) < len(candidates[j])
-	})
-	if len(candidates) == 0 {
-		return agentConsoleFallbackNonEscapeBind(target)
-	}
-	return candidates[0], true
-}
-
-func agentConsoleFallbackNonEscapeBind(target inputrc.Bind) (string, bool) {
-	switch target.Action {
-	case "previous-history", "history-search-backward":
-		return inputrc.Unescape(`\C-p`), true
-	case "next-history", "history-search-forward":
-		return inputrc.Unescape(`\C-n`), true
-	case "backward-char", "vi-backward-char":
-		return inputrc.Unescape(`\C-b`), true
-	case "forward-char", "vi-forward-char":
-		return inputrc.Unescape(`\C-f`), true
-	case "beginning-of-line":
-		return inputrc.Unescape(`\C-a`), true
-	case "end-of-line":
-		return inputrc.Unescape(`\C-e`), true
-	default:
-		return "", false
-	}
-}
-
-func agentConsoleReadlineSequence(seq string) string {
-	if seq == "" {
-		return ""
-	}
-	converted := make([]rune, 0, len(seq))
-	for _, r := range seq {
-		if inputrc.IsMeta(r) {
-			converted = append(converted, inputrc.Esc, inputrc.Demeta(r))
-			continue
-		}
-		converted = append(converted, r)
-	}
-	return string(converted)
 }
 
 func (r *AgentConsole) Start() error {
@@ -512,201 +290,6 @@ func (r *AgentConsole) executeArgs(ctx context.Context, args []string) error {
 	root.SetArgs(args)
 	root.SetContext(ctx)
 	return root.Execute()
-}
-
-// renderBanner prints a compact welcome block to stderr: title/version,
-// resolved model, the session mode, and a short next-step hint. It uses fixed
-// ANSI tokens so redirected or recorded sessions do not receive terminal
-// background probes. stderr-TTY-only and skipped in quiet mode so redirected
-// logs stay clean. Printed once into the scrollback (PTY-forward safe).
-func (r *AgentConsole) renderBanner() {
-	if r.output == nil || r.output.VerbosityLevel() < 0 || r.output.Stderr() == nil {
-		return
-	}
-	if !r.output.tty {
-		return
-	}
-	fmt.Fprint(r.output.Stderr(), r.bannerOutput())
-}
-
-func (r *AgentConsole) bannerOutput() string {
-	colorEnabled := r.output != nil && r.output.color.Enabled
-	provider, model := r.providerModel()
-	modelText := "not configured - run `aiscan --init`"
-	modelStyle := ansiWarn
-	switch {
-	case provider != "" && model != "":
-		modelText = provider + " / " + model
-		modelStyle = ansiAccent
-	case provider != "":
-		modelText = provider
-		modelStyle = ansiAccent
-	}
-
-	width := r.bannerWidth()
-	header := ansiTitle("aiscan", colorEnabled) + " " + ansiDim("v"+cfg.Version, colorEnabled)
-
-	var lines []string
-	lines = append(lines, header)
-	lines = append(lines, bannerKV("model", modelStyle(modelText, colorEnabled), colorEnabled))
-	lines = append(lines, bannerKV("mode", ansiDim(r.sessionSummary(), colorEnabled), colorEnabled))
-	lines = append(lines, bannerKV("help", renderInlineCommands([]string{"/help", "/status", "/exit"}, colorEnabled), colorEnabled))
-	lines = append(lines, bannerKV("keys", ansiDim("Esc", colorEnabled)+" "+ansiDim("interrupt", colorEnabled)+ansiDim("  ", colorEnabled)+ansiDim("Ctrl+O", colorEnabled)+" "+ansiDim("verbosity", colorEnabled), colorEnabled))
-
-	box := renderFixedBox(strings.Join(lines, "\n"), width, colorEnabled)
-	intent := ansiDim("输入目标或任务即可；例如：扫描 192.168.1.10 的 Web 风险", colorEnabled)
-
-	var b strings.Builder
-	fmt.Fprintln(&b)
-	fmt.Fprintln(&b, box)
-	fmt.Fprintln(&b, "  "+intent)
-	if notice := strings.TrimSpace(r.startupNotice); notice != "" {
-		fmt.Fprintln(&b, "  "+ansiWarn("⚠ "+notice, colorEnabled))
-	}
-	fmt.Fprintln(&b)
-	return b.String()
-}
-
-func (r *AgentConsole) bannerWidth() int {
-	const (
-		minWidth     = 44
-		defaultWidth = 64
-		maxWidth     = 78
-	)
-	width := defaultWidth
-	if r != nil && r.terminal != nil && r.terminal.Control != nil {
-		if columns, _ := r.terminal.Control.Size(); columns > 0 {
-			width = columns - 4
-		}
-	} else if r != nil && r.output != nil && r.output.Stderr() != nil {
-		if columns := writerTerminalWidth(r.output.Stderr()); columns > 0 {
-			width = columns - 4
-		}
-	}
-	if width < minWidth {
-		return minWidth
-	}
-	if width > maxWidth {
-		return maxWidth
-	}
-	return width
-}
-
-func writerTerminalWidth(w io.Writer) int {
-	file, ok := w.(*os.File)
-	if !ok {
-		return 0
-	}
-	width, _, err := term.GetSize(int(file.Fd()))
-	if err != nil || width <= 0 {
-		return 0
-	}
-	return width
-}
-
-func bannerKV(label, value string, colorEnabled bool) string {
-	return ansiDim(fmt.Sprintf("%-9s", label), colorEnabled) + value
-}
-
-func renderFixedBox(body string, width int, colorEnabled bool) string {
-	const minInnerWidth = 16
-	innerWidth := width - 4
-	if innerWidth < minInnerWidth {
-		innerWidth = minInnerWidth
-	}
-	lines := strings.Split(body, "\n")
-	for _, line := range lines {
-		if n := visibleRuneLen(line); n > innerWidth {
-			innerWidth = n
-		}
-	}
-
-	border := func(s string) string { return ansiDim(s, colorEnabled) }
-	var b strings.Builder
-	fmt.Fprintf(&b, "%s\n", border("╭"+strings.Repeat("─", innerWidth+2)+"╮"))
-	for _, line := range lines {
-		padding := innerWidth - visibleRuneLen(line)
-		if padding < 0 {
-			padding = 0
-		}
-		fmt.Fprintf(&b, "%s %s%s %s\n",
-			border("│"),
-			line,
-			strings.Repeat(" ", padding),
-			border("│"))
-	}
-	fmt.Fprint(&b, border("╰"+strings.Repeat("─", innerWidth+2)+"╯"))
-	return b.String()
-}
-
-func visibleRuneLen(s string) int {
-	return len([]rune(outputpkg.StripANSI(s)))
-}
-
-func ansiWrap(s, code string, enabled bool) string {
-	return outputpkg.NewColor(enabled).Wrap(s, code)
-}
-
-func ansiTitle(s string, enabled bool) string {
-	return ansiWrap(s, outputpkg.ANSIBold+outputpkg.ANSICyan, enabled)
-}
-
-func ansiAccent(s string, enabled bool) string {
-	return ansiWrap(s, outputpkg.ANSICyan, enabled)
-}
-
-func ansiWarn(s string, enabled bool) string {
-	return ansiWrap(s, outputpkg.ANSIYellow, enabled)
-}
-
-func ansiDim(s string, enabled bool) string {
-	return ansiWrap(s, outputpkg.ANSIDim, enabled)
-}
-
-func renderInlineCommands(commands []string, colorEnabled bool) string {
-	parts := make([]string, 0, len(commands))
-	for _, command := range commands {
-		parts = append(parts, ansiAccent(command, colorEnabled))
-	}
-	return strings.Join(parts, ansiDim("  ", colorEnabled))
-}
-
-func (r *AgentConsole) sessionSummary() string {
-	var parts []string
-	if r != nil && r.output != nil {
-		switch r.output.mode {
-		case ModeStatic:
-			parts = append(parts, "static")
-		case ModeForwarded:
-			parts = append(parts, "forwarded")
-		default:
-			parts = append(parts, "pty")
-		}
-		if r.output.stream.enabled {
-			parts = append(parts, "stream")
-		} else if r.output.Markdown() {
-			parts = append(parts, "pretty")
-		} else {
-			parts = append(parts, "plain")
-		}
-	}
-	if r != nil && r.option != nil {
-		if space := strings.TrimSpace(r.option.Space); space != "" {
-			parts = append(parts, "space "+space)
-		}
-	}
-	if len(parts) == 0 {
-		return "pty"
-	}
-	return strings.Join(parts, " · ")
-}
-
-func (r *AgentConsole) providerModel() (string, string) {
-	if r.appInfo.Commands == nil {
-		return "", ""
-	}
-	pc := r.appInfo.ProviderConfig
-	return pc.Provider, pc.Model
 }
 
 func (r *AgentConsole) replSession() *Session {
@@ -983,72 +566,6 @@ func wrapCommand(cmd Command, s *Session) *cobra.Command {
 		}
 	}
 	return cc
-}
-
-// ---------------------------------------------------------------------------
-// Rendering: help, status, panels
-// ---------------------------------------------------------------------------
-
-func (r *AgentConsole) renderHelp() string {
-	colorEnabled := r.output != nil && r.output.color.Enabled
-	cmds := r.allCommands()
-	rows := make([]helpRow, 0, len(cmds)+3)
-	for _, c := range cmds {
-		if c.Hidden {
-			continue
-		}
-		rows = append(rows, helpRow{Command: c.Name, Detail: c.Description})
-	}
-	rows = append(rows, helpRow{})
-	rows = append(rows, helpRow{Command: "普通文本", Detail: "直接发送自然语言任务"})
-	rows = append(rows, helpRow{Command: "! <命令>", Detail: "直接执行 bash/伪命令（跳过 LLM）"})
-	return r.renderPanel("commands", renderHelpRows(rows, colorEnabled), colorEnabled)
-}
-
-func (r *AgentConsole) renderStatus() string {
-	colorEnabled := r.output != nil && r.output.color.Enabled
-	info := CollectStatus(r.replSession(), r.sessionSummary(), agentConsoleHistoryPath())
-	rows := []helpRow{
-		{Command: "model", Detail: info.Provider + " / " + info.Model},
-		{Command: "render", Detail: info.Mode},
-		{Command: "task", Detail: info.Task},
-		{Command: "ioa", Detail: info.IOA},
-		{Command: "history", Detail: info.History},
-	}
-	if info.Skills != "" {
-		rows = append(rows, helpRow{Command: "skills", Detail: info.Skills})
-	}
-	return r.renderPanel("status", renderHelpRows(rows, colorEnabled), colorEnabled)
-}
-
-type helpRow struct {
-	Command string
-	Detail  string
-}
-
-const helpRowCommandWidth = 18
-
-func renderHelpRows(rows []helpRow, colorEnabled bool) string {
-	var b strings.Builder
-	for _, row := range rows {
-		if row.Command == "" && row.Detail == "" {
-			b.WriteByte('\n')
-			continue
-		}
-		command := ansiAccent(fmt.Sprintf("%-*s", helpRowCommandWidth, row.Command), colorEnabled)
-		detail := ansiDim(row.Detail, colorEnabled)
-		fmt.Fprintf(&b, "%s%s\n", command, detail)
-	}
-	return strings.TrimRight(b.String(), "\n")
-}
-
-func (r *AgentConsole) renderPanel(title, body string, colorEnabled bool) string {
-	title = strings.TrimSpace(title)
-	if title == "" {
-		title = "aiscan"
-	}
-	header := ansiTitle(title, colorEnabled)
-	return "\n" + renderFixedBox(header+"\n"+body, r.bannerWidth(), colorEnabled) + "\n\n"
 }
 
 func (r *AgentConsole) ensureOutput() *AgentOutput {
